requirements.txt
charset-normalizer==3.0.1
idna==3.4
nba-api==1.1.14
numpy==1.24.2
oauthlib==3.2.2
python-dateutil==2.8.2
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.12.1
urllib3==1.26.14
